Expense CRUD

+++ expensemanager/src/main/java/com/baji/expensemanager/dto/ExpenseRequest.java
@@ -2,15 +2,16 @@ package com.baji.expensemanager.dto;
 
 import java.time.LocalDate;
 
+import com.fasterxml.jackson.annotation.JsonCreator;
+
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
-
 public record ExpenseRequest(
 		@NotBlank(message="Description is Required") String description,
 		@NotNull(message="Amount is Required")
 		@Positive(message="Amount must be greater than zero")
-		Long amount,
+		Double amount,
 		@NotBlank(message="Category is Required")
 		String category,
 		@NotNull(message = "Expense date is required")
